requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
